pyproject.toml
[project]
name = "nl-voting-data-scraper"
version = "0.1.0"
description = "Scrape Dutch voting advice (StemWijzer) data for any election"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Rehan Fazal" }]
keywords = ["stemwijzer", "dutch", "elections", "scraper", "votematch"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]
dependencies = [
    "httpx>=0.28",
    "pydantic>=2.12",
    "click>=8.3",
    "rich>=14.0",
    "pycryptodome>=3.23",
    "tenacity>=9.1",
]

[project.optional-dependencies]
browser = ["playwright>=1.58"]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "pytest-httpx>=0.36",
    "ruff>=0.15",
]

[project.scripts]
nl-voting-data-scraper = "nl_voting_data_scraper.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nl_voting_data_scraper"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
